blog-series/01-intro/requirements.txt
# Ollama dependencies
strands-agents[ollama]
langchain-ollama

# Bedrock dependencies
langchain-aws

# OpenAI dependencies
strands-agents[openai]
langchain-openai

# Shared dependencies
langgraph
langchain
